refactor(aapd-1081): clean up questionnaire details rows and formatting
